aurora-backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10

# PandasAI
pandasai==2.0.12

# Data Processing
# Note: pandas version is constrained by pandasai dependencies
pandas>=1.5.3,<2.0.0  # Compatible with pandasai 2.0.12
numpy>=1.24.0,<2.0.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# CORS is handled by FastAPI middleware (no separate package needed)

